requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.4
google-generativeai==0.8.6
google-api-python-client==2.134.0
supabase==2.5.1
httpx==0.27.0
python-multipart==0.0.9
celery==5.4.0
redis==5.0.6
boto3==1.34.144
whisper